requirements.txt
Flask==3.1.*
flask-cors==5.*
flask-jwt-extended==4.*
python-dotenv==1.*
bcrypt==4.*
influxdb-client==1.*
Flask-SQLAlchemy==3.*
Flask-Migrate==4.*
psycopg2-binary==2.*
